# Lab-2/api/openapi.yaml
openapi: 3.0.3

info:
  title: Film Manager
  version: 1.0.11
tags:
  - name: films
  - name: users
  - name: images

paths:
  # Root API  
  /api/:
    get:
      tags:
        - films
      summary: top level API endpoint
      description: it returns you a list with some useful service endpoints URLs
      operationId: 
        root
      responses:
        "200":
          description: some useful URLs
          content:
            application/json:
              schema:
                type: object
                properties:
                  links:
                    type: array
                    items:
                      type: object
                      properties:
                        rel:
                          type: string
                        href:
                          type: string

# Private films getters APIs
  /api/films/private:
    get:
      tags:
        - films
      summary:     fetch all films marked as private
      operationId: privateFilms
      responses:
        '200':
          description: private films in the database
          # Since the list of private films may be huge,
          # please adopt the pagination, i.e. returning
          # a limited number of results
          content:
            application/json:
              schema:
                type: object
                properties:
                  metatadata:
                    type: object
                    properties:
                      total: # How many items the database has stored
                       type: integer
                      count: # How many items are returned in the response
                        type: integer 
                      limit: # How many items the client wants to see
                        type: integer
                      offset: # From which item the client wants to see the items
                        type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Film'
  /api/films/private/{fid}:
    get:
      tags:
        - films
      summary: fetch a particular film
      operationId: privateFilm
      parameters:
        - name: fid
          in: path
          description: Film ID (the one to return)
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The requested film
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Film'  
        '401':
          description: authorization information is missing or invalid
          content:
            text/plain:
              schema:
                type: string        
        '404':
          description: film not found
  /api/films/private/{fid}/reviews:
    get:
      tags:
        - films
      summary:     fetch all reviews associated to a film marked as private
      operationId: reviewsPrivateFilm
      parameters:
        - name: fid
          in: path
          description: film ID (the one whose whose reviews are requested)
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: reviews associated to that film in the database
          content:
            application/json:
              # Since the list of reviews may be huge,
              # please adopt the pagination, i.e. returning
              # a limited number of results
              schema:
                type: object
                properties:
                  metatadata:
                    type: object
                    properties:
                      total: # How many items the database has stored
                       type: integer
                      count: # How many items are returned in the response
                        type: integer 
                      limit: # How many items the client wants to see
                        type: integer
                      offset: # From which item the client wants to see the items
                        type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'      
        '404':
         description: film not found    
  /api/films/private/{fid}/reviews/{uid}:
    get:
      tags:
        - films
      summary:     fetch a particular review for a film marked as private
      operationId: reviewPrivateFilm
      parameters:
        - name: fid
          in: path
          description: film ID (the one whose review is requested)
          required: true
          schema:
            type: integer
        - name: uid
          in: path
          description: user ID (the one whose review is requested)
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: the requested review
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Review'        
        '404':
         description: review not found

# Public films getters APIs
  /api/films/public:
    get:
      tags:
        - films
      summary:     fetch all films marked as public
      operationId: publicFilms
      responses:
        '200':
          description: public films in the database
          # Since the list of public films may be huge,
          # please adopt the pagination, i.e. returning
          # a limited number of results
          content:
            application/json:
              schema:
                type: object
                properties:
                  metatadata:
                    type: object
                    properties:
                      total: # How many items the database has stored
                       type: integer
                      count: # How many items are returned in the response
                        type: integer 
                      limit: # How many items the client wants to see
                        type: integer
                      offset: # From which item the client wants to see the items
                        type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Film'
  /api/films/public/{fid}:
    get:
      tags:
        - films
      summary:     fetch a particular film marked as public
      operationId: publicFilm
      parameters:
        - name: fid
          in: path
          description: film ID (the one to return)
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: public film in the database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Film'          
        '404':
          description: film not found  
  /api/films/public/{fid}/reviews:
    get:
      tags:
        - films
      summary: fetch all reviews associated to a film marked as public
      operationId: reviewsPublicFilm
      parameters:
        - name: fid
          in: path
          description: film ID (the one whose whose reviews are requested)
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: reviews associated to that film in the database
          content:
            application/json:
              # Since the list of reviews may be huge,
              # please adopt the pagination, i.e. returning
              # a limited number of results
              schema:
                type: object
                properties:
                  metatadata:
                    type: object
                    properties:
                      total: # How many items the database has stored
                       type: integer
                      count: # How many items are returned in the response
                        type: integer 
                      limit: # How many items the client wants to see
                        type: integer
                      offset: # From which item the client wants to see the items
                        type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'       
        '404':
         description: film not found    
  /api/films/public/{fid}/reviews/{uid}:
    get:
      tags:
        - films
      summary:     fetch a particular review for a film marked as public
      operationId: reviewPublicFilm
      parameters:
        - name: fid
          in: path
          description: film ID (the one whose whose review is requested)
          required: true
          schema:
            type: integer
        - name: uid
          in: path
          description: user ID (the one whose whose review is requested)
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 
            The requested review
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Review'        
        '404':
         description: review not found

# Write and delete films APIs
  /api/films:
    post:
      tags:
        - films
      summary:     create a new film
      operationId: newFilm
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/New-Film-Request-Body'
      responses:
        '200':
          description: new film in database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Film'
        '401':
          description: authorization information is missing or invalid
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: bad request
          content:
            text/plain:
              schema:
                type: string
  /api/films/{fid}:
    put:
      tags:
        - films
      summary: edit a film
      operationId: editFilm
      parameters:
        - name: fid
          in: path
          description: film ID (the one to edit)
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Film'
      responses:
        '200':
          description: the updated film resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Film'          
        '204':
          description: content is null
        '400':
          description: bad request
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: authorization information is missing or invalid
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: film not found
    delete:
      tags:
        - films
      summary: delete a film
      operationId: deleteFilm
      parameters:
        - name: fid
          in: path
          description: film ID (the one to delete)
          required: true
          schema:
            type: string
      responses:
        '204':
          description: content is null
        '400':
          description: bad request
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: authorization information is missing or invalid
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Film not found

# High-level getters films APIs
  /api/films/created:
    get:
      tags:
        - films
      summary: fetch all films created the authenticated user has created
      operationId: createdFilms
      responses:
        '200':
          description: the requested films
          # Since the list of films may be huge,
          # please adopt the pagination, i.e. returning
          # a limited number of results
          content:
            application/json:
              schema:
                type: object
                properties:
                  metatadata:
                    type: object
                    properties:
                      total: # How many items the database has stored
                       type: integer
                      count: # How many items are returned in the response
                        type: integer 
                      limit: # How many items the client wants to see
                        type: integer
                      offset: # From which item the client wants to see the items
                        type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Film'             
  /api/films/invited-to-review:
    get:
      tags:
        - films
      summary: fetch all films to which the authenticated user has been invited to review
      operationId: filmsToReview
      responses:
        '200':
          description: 
            The requested films
          # Since the list of films may be huge,
          # please adopt the pagination, i.e. returning
          # a limited number of results
          content:
            application/json:
              schema:
                type: object
                properties:
                  metatadata:
                    type: object
                    properties:
                      total: # How many items the database has stored
                       type: integer
                      count: # How many items are returned in the response
                        type: integer 
                      limit: # How many items the client wants to see
                        type: integer
                      offset: # From which item the client wants to see the items
                        type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Film'    

# Users related APIs
  /api/user/auth:
    post:
      tags:
        - users
      summary: login the user
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login-Request-Body' 
      responses:
        '200':
          description: user sucessfully logged-in
        '400':
          description: invalid username/password supplied
  /api/user/no-auth:
    post:
      tags:
        - users
      summary: logout the user
      operationId: logout
      responses:
        '200':
          description: user sucessfully logged-out

# Reviews related APIs
  /api/reviews:
    post:
      tags:
        - reviews
      summary: create a new review
      operationId: newReview
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
      responses:
        '200':
          description: new review in database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '401':
          description: authorization information is missing or invalid
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: bad request
          content:
            text/plain:
              schema:
                type: string  
  /api/reviews/{fid}/{uid}:
    put:
      tags:
        - reviews
      summary: edit a review
      operationId: 
        editReview
      parameters:
        - name: fid
          in: path
          description: film ID (the one whose review is requested)
          required: true
          schema:
            type: integer
        - name: uid
          in: path
          description: owner ID (the one to edit)
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
      responses:
        '200':
          description: the updated review resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'          
        '204':
          description: content is null
        '400':
          description: bad request
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: authorization information is missing or invalid
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: film not found
    delete:
      tags:
        - reviews
      summary: 
        Delete a review
      operationId: 
        deleteReview
      parameters:
        - name: fid
          in: path
          description: film ID (the one whose review is requested)
          required: true
          schema:
            type: integer
        - name: uid
          in: path
          description: owner ID (the one whose to delete)
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: content is null
        '400':
          description: bad request
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: authorization information is missing or invalid
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Film not found

# Images related APIs
  /api/films/{fid}/images:
    get:
      tags:
        - images
      summary: fetch all images related to a film
      operationId: filmImages
      parameters:
        - name: fid
          in: path
          description:  film ID (the one whose whose images are requested)
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: the requested images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Image"
        '401':
          description: authorization information is missing or invalid
          content:
            text/plain:
              schema:
                type: string        
        '404':
          description: film not found
    post:
      tags:
        - images
      summary: upload a new image
      operationId: newImage
      parameters:
        - name: fid
          in: path
          description: film ID (the one to which associate the film)
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/New-Image-Request-Body'
      responses:
        '200':
          description: new image in database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        '401':
          description: authorization information is missing or invalid
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: 
            Bad request
          content:
            text/plain:
              schema:
                type: string
  /api/films/{fid}/images/{iid}:
    get:
      tags:
        - images
      summary: fetch an image related to a film
      operationId: filmImage
      parameters:
        - name: fid
          in: path
          description:  film ID (the one whose whose image are requested)
          required: true
          schema:
            type: integer
        - name: iid
          in: path
          description:  image ID (the one which is requested)
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: the requested image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Image"
            image/png:
              schema: 
                  type:   string
                  format: base64 
            image/jpeg:
              schema: 
                  type:   string
                  format: base64
            image/gif:
              schema: 
                  type:   string
                  format: base64       
        '401':
          description: authorization information is missing or invalid
          content:
            text/plain:
              schema:
                type: string        
        '404':
          description: film not found
          
    delete:
      tags:
        - images
      summary: delete an image
      operationId: deleteImage
      parameters:
        - name: fid
          in: path
          description: film ID (the one whose review is requested)
          required: true
          schema:
            type: integer
        - name: iid
          in: path
          description: image ID (the one to delete)
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: content is null
        '400':
          description: bad request
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: authorization information is missing or invalid
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: film not found
          
components:
  schemas:
    Film:
      type: object
      properties:
        fid:
          type: integer
        title:
          type: string
        owner:
          type: integer
        private:
          type: integer
        watchDate:
          type:   string
          format: date
        rating:
          type: integer
          minimum: 0
          maximum: 10
        favorite:
          type:    integer
          default: false
      required:
        - fid
        - owner
        - private
    Review:
      type: object
      properties:
        fid:
          type: integer
        uid:
          type: integer
        completed:
          type: integer
        reviewDate:
          type: string
          format: date
        rating:
          type: integer
          minimum: 0
          maximum: 10
        text:
          type: string
          maxLength: 1000
      required:
        - fid
        - uid
        - completed
    User:
      type: object
      properties:
        uid:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
          maxLength: 20
      required:
        - uid
        - email
        - password
    Image:
      type: object
      properties:
        fid:
          type: integer
        name:
          type: string
      required:
        - fid
    New-Image-Request-Body:
      type: object
      properties:
        name:
          type: string
        data:
          type: object
          properties:
            format:
              type: string
            bytes:
              type:   string
              format: base64
    Login-Request-Body:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    New-Film-Request-Body:
      type: object
      properties:
        title:
          type: string
        owner:
          type: integer
        private:
          type: integer
        watchDate:
          type: string
          format: date
        rating:
          type: integer
          minimum: 0
          maximum: 10
        favorite: 
          type: integer
          default: false
      required:
        - owner
        - private